% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_clipboard.R
\name{abs_clipboard}
\alias{abs_clipboard}
\title{Assign a really long URL to a variable}
\usage{
abs_clipboard()
}
\value{
An extremely long character string (4087+ chars) in its entirety
}
\description{
The RStudio console can only take string inputs with less than 4087
characters. This function gets around this by assigning the last string
you copied to a variable.
}
\examples{
\dontrun{
### Once you have copied the URL, call this function and assign it to a
### variable name, which will be passed through the \code{read_abs_api}
### function:

url <- abs_clipboard()
read_abs_api(url)
}
}
