% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_join.R
\name{api_join}
\alias{api_join}
\title{A special ABS join for the API}
\usage{
api_join(.x, col, dataset)
}
\arguments{
\item{.x}{(data.frame) A data frame in the dimensions list from
\code{tidy_api_data}.}

\item{col}{(column) The column from the dataset tibble from
\code{tidy_api_data} that corresponds to the data frame used.}

\item{dataset}{(tibble) The dataset tibble from \code{tidy_api_data}, which
has integers as the variable values/dimension members.}
}
\description{
This function renames the id column to the variable/dimension name for a data
frame in the dimensions list from \code{tidy_api_data}. It also adds a column
to the data frame, which contains all the row numbers.
}
