% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_split_url.R
\name{api_split_url}
\alias{api_split_url}
\title{Splits one large API URL into smaller URLs}
\usage{
api_split_url(url, old_api)
}
\arguments{
\item{url}{(char) The API URL used to query the data you want from the ABS.
Both the ABS.Stat URL structure and the new API URL structure (as outlined
on api.gov.au) are accepted. The following parameters would just need to be
included in the URLs to organise the data in the flat view and JSON format:
dimensionAtObservation=allDimensions (for the ABS.Stat SDMX-JSON API) and
format=jsondata&dimensionAtObservation=AllDimensions (for the new API).}

\item{old_api}{(boolean) States whether the ABS' API uses the old (ABS.Stat)
or new structure. If \code{old_api} == TRUE, then the API follows the old
structure.}
}
\value{
(character) A character vector with the smaller URLs.
}
\description{
This function splits up large URLs (URLs with more than 1000 characters) into
smaller batches to increase efficiency. It identifies the large variables
(dimensions) in the data key (those that have more than 50 values) and
divides these up into batches of 50 values. Smaller URLs are
then constructed using these batches.
}
