% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lfs_grossflows.R
\name{check_lfs_grossflows}
\alias{check_lfs_grossflows}
\title{Internal function to check if the data frame returned by read_lfs_grossflows()
contains expected unique values in key columns}
\usage{
check_lfs_grossflows(df)
}
\arguments{
\item{df}{data frame containing gross flows data}
}
\description{
Internal function to check if the data frame returned by read_lfs_grossflows()
contains expected unique values in key columns
}
\keyword{internal}
