% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chunk_api_url.R
\name{chunk_big_dim}
\alias{chunk_big_dim}
\title{Chunk a \code{query_url} into requests for fewer dimensions}
\usage{
chunk_big_dim(query_url, n = 300)
}
\arguments{
\item{query_url}{(character) the 'Data query' URL from the 'Developer API'
section from the \href{https://explore.data.abs.gov.au/}{ABS' API}. I'm not
sure what the difference between 'Flat' and 'Time series' is, because even
though the URLs are different they return the same data.}

\item{n}{(\code{numeric}; default = \code{300}) the number of dimension
levels to request in each URL. Higher means fewer URLs (and thus faster),
but increases the risk of the resulting URLs not being valid. Cursory
testing indicates that more than 400 results in data loss (ie some URLs are
still too big).}
}
\value{
(\code{character} vector) a vector of URLs, one for subgroup of the
  largest dimension.
}
\description{
One way to break up a \code{query_url} is by requesting fewer dimensions
(such as only a few SA2s as a time). This function does that for the
dimension with the most levels.
}
