% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chunk_api_url.R
\name{chunk_query_url}
\alias{chunk_query_url}
\title{Break up a \code{query_url} into chunks, by date and dimension}
\usage{
chunk_query_url(query_url, n = 75)
}
\arguments{
\item{query_url}{(character) the 'Data query' URL from the 'Developer API'
section from the \href{https://explore.data.abs.gov.au/}{ABS' API}. I'm not
sure what the difference between 'Flat' and 'Time series' is, because even
though the URLs are different they return the same data.}

\item{n}{(\code{numeric}; default = \code{300}) the number of dimension
levels to request in each URL. Higher means fewer URLs (and thus faster),
but increases the risk of the resulting URLs not being valid. Cursory
testing indicates that more than 400 results in data loss (ie some URLs are
still too big).}
}
\value{
(\code{character} vector) a \code{character} vector of URLs for the
  ABS API
}
\description{
Some URLs provided by the ABS' API are too long to actually work. This
function breaks up a \code{query_url} into multiple smaller ones. It does
this by looking at the time period the request covers, and separating it into
individual years, and by looking at the dimension with the most levels
(usually a geography) and creating URLs that only request a smaller number of
levels.
}
\examples{
\dontrun{
# A simple URL

# Labour force data query URL, taken from ABS API site
# (In `paste0()` to prevent help docs being truncated)
lf_url <- paste0("https://api.data.abs.gov.au/data/ABS,LF,1.0.0/",
                 "M9.3.1599.30+10+20.AUS.M?startPeriod=2021-01&",
                 "endPeriod=2021-12&dimensionAtObservation=AllDimensions")

lf <- read_abs_api(lf_url)



# A URL that might need breaking up (note the different start date)
# (this URL doesn't actually need breaking up but it is a good example)

lf_url2 <- paste0("https://api.data.abs.gov.au/data/ABS,LF,1.0.0/",
                  "M9.3.1599.30+10+20.AUS.M?startPeriod=1978-01&",
                  "endPeriod=2021-12&dimensionAtObservation=AllDimensions")

lf_urls <- chunk_query_url(lf_url2)
read_api_safely <- purrr::safely(read_abs_api)
lf_raw <- purrr::map(lf_urls, read_api_safely)
lf_data <- purrr::map_dfr(lf_raw, "result")
}
}
