% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_chunk_api_url.R
\name{chunk_span}
\alias{chunk_span}
\title{Chunk a \code{query_url} into individual years}
\usage{
chunk_span(query_url)
}
\arguments{
\item{query_url}{(character) the 'Data query' URL from the 'Developer API'
section from the \href{https://explore.data.abs.gov.au/}{ABS' API}. I'm not
sure what the difference between 'Flat' and 'Time series' is, because even
though the URLs are different they return the same data.}
}
\value{
(\code{character} vector) a vector of URLs, one for each year of data
  requested.
}
\description{
One way to break up a \code{query_url} is by year. This function does that.
}
