% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_by_region.r
\name{data_by_region}
\alias{data_by_region}
\title{Get the Data by Region URL}
\usage{
data_by_region(start = NULL, end = NULL)
}
\arguments{
\item{start}{(\code{numeric}; default = \code{NULL})}

\item{end}{(\code{numeric}; default = \code{NULL})}
}
\value{
(\code{character}) a query URL for Data by Region
}
\description{
This URL is such a pain to get into R, I thought it would be easier to just write a helper function. It reads the 25,000 character Data by Region URL from a text file into R, and allows the user to set the start and end years. By default it will just set to last year.
}
