% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fast_str_squish}
\alias{fast_str_squish}
\title{Very slightly faster version of stringr's str_squish()}
\usage{
fast_str_squish(string)
}
\arguments{
\item{string}{A string to squish (remove whitespace)}
}
\description{
Very slightly faster version of stringr's str_squish()
}
\keyword{internal}
