% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_abs.R
\name{find_abs}
\alias{find_abs}
\title{Find ABS time series of interest}
\usage{
find_abs(cat_no = NULL, cat_desc = NULL, find = NULL)
}
\arguments{
\item{cat_no}{ABS catalogue number, such as "6202.0".}

\item{cat_desc}{Description of an ABS product, such as "Labour Force". If `cat_no` isn't
specified, `cat_desc` must be specified. If `cat_no` is specified, `cat_desc` is ignored.}

\item{find}{String(s) to search for, such as "unemployment" or c("unemployed", "males", "New South Wales")}
}
\value{
A data frame containing information about time series that match your search term(s).
}
\description{
\code{find_abs()} looks in the ABS Time Series Directory for series
that match search terms you provide.
}
\examples{

\donttest{
 wpi <- find_abs(cat_no = "6345.0", find = c("private", "New South Wales", "excluding bonuses"))

}

}
