% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_cube.r
\name{get_available_lfs_cubes}
\alias{get_available_lfs_cubes}
\title{Show the available Labour Force, Australia, detailed data cubes that can be
downloaded}
\usage{
get_available_lfs_cubes()
}
\description{
Show the available Labour Force, Australia, detailed data cubes that can be
downloaded
}
\details{
Intended to be used with \code{read_lfs_datacube()}. Call
\code{read_lfs_datacube()} interactively, find the table of interest
(eg. "LM1"), then use \code{read_lfs_datacube()}.
}
\examples{

get_available_lfs_cubes()

}
