% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_api.R
\name{guess_structure_url}
\alias{guess_structure_url}
\title{Guess the structure URL for the API}
\usage{
guess_structure_url(query_url)
}
\arguments{
\item{query_url}{(character) the}
}
\value{
a URL string
}
\description{
The structure URL will return information about how certain variables are
encoded. This function can guess what it should be, given a query URL.
}
