% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_api.R
\name{label_abs_api_data}
\alias{label_abs_api_data}
\title{Label read_api data with datastructure metadata}
\usage{
label_abs_api_data(read_api_data, read_api_datastructure_data)
}
\arguments{
\item{read_api_data}{A data.frame from \code{\link{read_api}}.}

\item{read_api_datastructure_data}{The corresponding data.frame from \code{\link{read_api_datastructure}}. Dataflow id should be identical to \code{read_api_data}.}
}
\description{
Creates additional columns in the data.frame generated from \code{\link{read_api}} corresponding to metadata columns from \code{\link{read_api_datastructure}}.
}
\examples{
## Not run:

# Api data downloaded separately
abs_api_cpi <- read_api("CPI")
abs_api_ds_cpi <- read_api_datastructure("CPI")

label_abs_api_data(abs_api_cpi, abs_api_ds_cpi)

# Combined functions
label_abs_api_data(read_api("CPI"), read_api_datastructure("CPI"))

## End (Not run)
}
