% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read_abs_api.R
\name{open_sdmx}
\alias{open_sdmx}
\title{Wrapper to download and read an SDMX XML}
\usage{
open_sdmx(query_url)
}
\arguments{
\item{query_url}{(character) the 'Data query' URL from the 'Developer API'
section from the \href{https://explore.data.abs.gov.au/}{ABS' API}. I'm not
sure what the difference between 'Flat' and 'Time series' is, because even
though the URLs are different they return the same data.}
}
\value{
a \code{tibble}
}
\description{
There are two packages to read SDMX APIs in R: \code{readsdmx} and
\code{rsdmx}. The former is much faster, but hasn't been updated since 2019
and sometimes throws weird errors. The latter is works reliably, but is much,
much slower. Anyway, this function tries to improve the stability of the
former by handling the download explicitly.
}
