% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_query_url_helpers.R
\name{qu_biggest_dimesion}
\alias{qu_biggest_dimesion}
\title{What is the biggest dimension component in a \code{query_url}?}
\usage{
qu_biggest_dimesion(query_url)
}
\arguments{
\item{query_url}{(character) the 'Data query' URL from the 'Developer API'
section from the \href{https://explore.data.abs.gov.au/}{ABS' API}. I'm not
sure what the difference between 'Flat' and 'Time series' is, because even
though the URLs are different they return the same data.}
}
\value{
(\code{character}) the longest part of \code{query_url} between two
  \code{.}.
}
\description{
This function returns the biggest dimension (ie the most verbose, so often
geography) in a \code{query_url}. This is helpful for chunking up a long
\code{query_url} into smaller requests.
}
