% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_query_url_helpers.R
\name{qu_has_span}
\alias{qu_has_span}
\title{Check if a \code{query_url} has specified start and end periods}
\usage{
qu_has_span(query_url)
}
\arguments{
\item{query_url}{(character) the 'Data query' URL from the 'Developer API'
section from the \href{https://explore.data.abs.gov.au/}{ABS' API}. I'm not
sure what the difference between 'Flat' and 'Time series' is, because even
though the URLs are different they return the same data.}
}
\value{
(\code{logical}) \code{TRUE} if \code{query_url} has
  \code{startPeriod} and \code{endPeriod}
}
\description{
This is helpful for chunking up a long \code{query_url} into smaller
requests.
}
