% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_api.R
\name{qu_ok}
\alias{qu_ok}
\title{Superficial check if a \code{query_url} is reachable}
\usage{
qu_ok(query_url)
}
\arguments{
\item{query_url}{(character) the 'Data query' URL from the 'Developer API'
section from the \href{https://explore.data.abs.gov.au/}{ABS' API}. I'm not
sure what the difference between 'Flat' and 'Time series' is, because even
though the URLs are different they return the same data.}
}
\value{
(\code{logical}) \code{TRUE} if the status code is 200, \code{FALSE}
  otherwise (with an error message printed to the console)
}
\description{
Some URLs provided by the ABS' API are actually duds, so it can help to check
if they can be reached. This does that by calling \code{httr::GET()},
checking the status response, and printing a (hopefully) helpful error
message.
}
