% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_codebook.R
\name{read_abs_codebook}
\alias{read_abs_codebook}
\title{Extracts ABS-formatted codebook directly from Excel spreadsheets.}
\usage{
read_abs_codebook(path, sheet)
}
\arguments{
\item{path}{Filepath to Excel spreadsheet.}

\item{sheet}{Sheet name or number.}
}
\value{
dataframe
}
\description{
Extracts ABS-formatted codebook directly from Excel spreadsheets.
}
