% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_local.R
\name{read_abs_local}
\alias{read_abs_local}
\title{Read and tidy locally-saved ABS time series spreadsheet(s)}
\usage{
read_abs_local(filenames = NULL, path = "data/ABS")
}
\arguments{
\item{filenames}{character vector of at least one filename of a locally-stored
ABS time series spreadsheet. For example, "6202001.xls" or c("6202001.xls", "6202005.xls").
If `filenames` is blank, `read_abs_local()` will attempt to read all .xls files in
the directory specified with `path`. Please note that the directory should
be specified with `path`.}

\item{path}{path to local directory containing ABS time series file(s).
Default is "data/ABS". If nothing is specified in `filenames`,
`read_abs_local()` will attempt to read all .xls files in the directory specified with `path`.}
}
\description{
If you need to download and tidy time series data from the ABS, use \code{read_abs()}.
\code{read_abs_local()} imports and tidies data from ABS time series spreadsheets that
are already saved to your local drive.
}
\details{
Unlike \code{read_abs()}, the `table_title` column in the data frame
returned by \code{read_abs_local()} is blank. If you require `table_title`,
please use \code{read_abs()} instead.
}
\examples{

 # Load and tidy two specified files from the "data/ABS" subdirectory
 # of your working directory

 \donttest{lfs <- read_abs_local(c("6202001.xls", "6202005.xls"))}


}
