% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_data.R
\name{read_abs_metadata}
\alias{read_abs_metadata}
\title{Extracts ABS series metadata directly from Excel spreadsheets and converts to long-form.}
\usage{
read_abs_metadata(path, sheet)
}
\arguments{
\item{path}{Filepath to Excel spreadsheet.}

\item{sheet}{Sheet name or number.}
}
\value{
Long-form dataframe
}
\description{
Extracts ABS series metadata directly from Excel spreadsheets and converts to long-form.
}
\examples{
\dontrun{
read_abs_metadata(path=system.file("extdata", "5206002_expenditure_volume_measures.xls",  package = "readabs", mustWork = TRUE), sheet=2)
}

}
