% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_sdmx.R
\name{read_abs_sdmx}
\alias{read_abs_sdmx}
\title{Extracts ABS XML-formatted data using the exported XML link at http://www.abs.gov.au/}
\usage{
read_abs_sdmx(url)
}
\arguments{
\item{url}{URL weblink.}
}
\value{
Dataframe
}
\description{
Extracts ABS XML-formatted data using the exported XML link at http://www.abs.gov.au/
}
\examples{
read_abs_sdmx(url="http://stat.data.abs.gov.au/restsdmx/sdmx.ashx/GetData/CPI/1.1+2+3+4+5+6+7+8+50.10001.10.Q/all?startTime=2014-Q1&endTime=2016-Q4")
}
