% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readabs-package.R
\docType{package}
\name{readabs-package}
\alias{readabs}
\alias{readabs-package}
\title{readabs: Download and Tidy Time Series Data from the Australian Bureau of Statistics}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Downloads, imports, and tidies time series data from the 
    Australian Bureau of Statistics <https://www.abs.gov.au/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattcowgill/readabs}
  \item Report bugs at \url{https://github.com/mattcowgill/readabs/issues}
}

}
\author{
\strong{Maintainer}: Matt Cowgill \email{mattcowgill@gmail.com} (\href{https://orcid.org/0000-0003-0422-3300}{ORCID})

Authors:
\itemize{
  \item Zoe Meers \email{zoe.meers@sydney.edu.au}
  \item Jaron Lee \email{jaron.lee@sydney.edu.au}
  \item David Diviny \email{David.Diviny@nousgroup.com.au}
}

Other contributors:
\itemize{
  \item Hugh Parsonage \email{hugh.parsonage@gmail.com} [contributor]
}

}
\keyword{internal}
