% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_files.R
\name{search_files}
\alias{search_files}
\title{Search for a file within an ABS catalogue}
\usage{
search_files(string, catalogue, refresh = FALSE)
}
\arguments{
\item{string}{String to search for among filenames in a catalogue}

\item{catalogue}{Name of catalogue}

\item{refresh}{logical; `FALSE` by default. When `TRUE`, will re-scrape the
list of files within the catalogue.}
}
\description{
Search for a file within an ABS catalogue
}
\examples{
\dontrun{  search_files("GM1", "labour-force-australia") }
}
