% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_catalogues.r
\name{show_available_catalogues}
\alias{show_available_catalogues}
\title{Helper function for \code{download_abs_data_cube} to show the available catalogues.}
\usage{
show_available_catalogues(selected_heading = NULL, refresh = FALSE)
}
\arguments{
\item{selected_heading}{optional character string specifying the heading on the \href{https://www.abs.gov.au/statistics}{ABS statistics webpage}.
e.g. "Earnings and work hours"}

\item{refresh}{logical; \code{FALSE} by default. If \code{FALSE}, an internal table of the available ABS catalogues is used. If \code{TRUE}, this table is refreshed from the ABS website.}
}
\value{
a character vector of catalogues.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function lists the possible catalogues that are available on the ABS website.
These catalogues must be specified as a string as an argument to \code{download_abs_data_cube}.
}
\examples{
show_available_catalogues("Earnings and work hours")
}
\seealso{
Other data cube functions: 
\code{\link{download_abs_data_cube}()},
\code{\link{search_catalogues}()},
\code{\link{show_available_files}()}
}
\concept{data cube functions}
