% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadABS.R
\name{tidy_ABS}
\alias{tidy_ABS}
\title{Tidy an xls or xlsx file}
\usage{
tidy_ABS(xl_workbook, allowed_blank_rows = 1, spreadsheet_type = "Data cube")
}
\arguments{
\item{xl_workbook}{Path to the input file}

\item{allowed_blank_rows}{Number of blank rows between tables, options are 1 or 2. Defaults to 1.}

\item{spreadsheet_type}{Spreadsheet type. Options are 'Time series', 'Data cube', 'Census'. Defaults to 'Data cube'.}
}
\value{
A list containing the cleaned data and the tab, row and column information on where the data was found
}
\description{
This function puts xls or xlsx files in a standardised 'long' format, retaining descriptive information in the spreadsheet. It is calibrated to work on Australian Bureau of Statistics, 'ABS' spreadsheets.
}
