% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tidy_api_data.R
\name{tidy_api_data}
\alias{tidy_api_data}
\title{Tidy the raw data returned using the ABS' API}
\usage{
tidy_api_data(query_data, structure_url)
}
\arguments{
\item{query_data}{(\code{data.frame}) raw data as a \code{data.frame} returned
from the ABS API}

\item{structure_url}{(character; default = \code{NULL}) the 'Structure query'
URL from the same place. This provides the information for the data
dictionary. If \code{NULL} it will be guessed from \code{query_url}.}
}
\value{
(tibble) Returns a tibble containing the ABS data you queried.
}
\description{
This function gets the nested list of raw data returned using the API URL and
transforms it into into a tidy tibble based on the structure of the API URL
(i.e. whether it follows the old or new structure, as stated in the
\code{old_api} parameter).
}
