% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_awe.R
\name{tidy_awe}
\alias{tidy_awe}
\title{Internal function to tidy a dataframe from ABS 6302}
\usage{
tidy_awe(df)
}
\arguments{
\item{df}{Data frame containing table 2 from ABS 6302, imported using `read_abs()`}
}
\description{
Internal function to tidy a dataframe from ABS 6302
}
\keyword{internal}
