% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_chars.R
\name{unique_chars}
\alias{unique_chars}
\title{Show the unique values of character columns}
\usage{
unique_chars(.data)
}
\arguments{
\item{.data}{a \code{tibble}}
}
\value{
a \code{list} where each element is a single column \code{tibble},
  referring to a character column of \code{.data}, and contains the unique
  values
}
\description{
A handy companion function for \code{read_abs_api}. Pass the results of an API
call to this and this will show you all the unique levels for all character
columns. Useful to inform filtering, plotting, etc.
}
\examples{
\dontrun{
new_url <- paste0("https://api.data.abs.gov.au/data/ALC/",
  "all?startPeriod=2010&endPeriod=2016&format=jsondata",
  "&dimensionAtObservation=AllDimensions")

tidy_data <- read_abs_api(new_url)
unique_chars(tidy_data)
}
}
