% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs.R
\name{get_abs}
\alias{get_abs}
\title{Download, extract, and tidy ABS time series spreadsheets (soft deprecated)}
\usage{
get_abs(...)
}
\arguments{
\item{...}{arguments passed to \code{read_abs()}}
}
\description{
\code{get_abs()} is deprecated. Please use \code{read_abs()} instead. It has identical functionality.
}
