% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_query_url_helpers.R
\name{qu_ok}
\alias{qu_ok}
\title{Superficial check if a \code{query_url} is reachable}
\usage{
qu_ok(res)
}
\arguments{
\item{res}{a HTTP response as returned by \code{httr::GET}}
}
\value{
(\code{logical}) \code{TRUE} if the status code is 200, \code{FALSE}
  otherwise (with an error message printed to the console)
}
\description{
Some URLs provided by the ABS' API are actually duds, so it can help to check
if they can be reached. This does that by calling \code{httr::GET()},
checking the status response, and printing a (hopefully) helpful error
message.
}
