% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read_abs_api.R
\name{read_abs_api}
\alias{read_abs_api}
\title{Get tidy data from ABS' API}
\usage{
read_abs_api(query_url, structure_url = NULL, verbose = TRUE, check = TRUE)
}
\arguments{
\item{query_url}{(character) the 'Data query' URL from the 'Developer API'
section from the \href{https://explore.data.abs.gov.au/}{ABS' API}. I'm not
sure what the difference between 'Flat' and 'Time series' is, because even
though the URLs are different they return the same data.}

\item{structure_url}{(character; default = \code{NULL}) the 'Structure query'
URL from the same place. This provides the information for the data
dictionary. If \code{NULL} it will be guessed from \code{query_url}.}

\item{verbose}{(logical; default = \code{TRUE}) print status update messages?}

\item{check}{(logical; default = \code{TRUE}) check the URL can be queried?
Some URLs are too long to be queried, so this can provide some advice on
how to deal with them (at the expense of speed).}
}
\value{
(\code{tibble}) Returns a tidy \code{tibble} containing the ABS data
  you queried. Each \code{tibble} returns a column \code{value} along with a
  series of pairs of other columns for the metadata (ear pair is suffixed
  \code{_code} and \code{_name}). The \code{_code} column shows the encoded
  value (e.g. \code{sex_code} might show a value of \code{3}), and the
  \code{_name} shows a human interpretable translation of \code{_code} (e.g.
  \code{sex_name} would show \code{Persons}).
}
\description{
This function returns a tidy \code{tibble} of data queried from the
\href{https://explore.data.abs.gov.au/}{ABS' API}. If the \code{query_url} is
too big the API to work (which happens sometimes), you may need to use this
function in conjunction with \code{\link{chunk_query_url}} and vectorisation
to query the API (see below for a simple example, with more information in
the vignette).
}
\details{
See the vignette for a detailed guide on how to use this to get monthly
labour force data or download the last decade of Data by Region at SA2.
}
\examples{
\dontrun{
# A simple URL

# Labour force data query URL, taken from ABS API site
# (In `paste0()` to prevent help docs being truncated)
lf_url <- paste0("https://api.data.abs.gov.au/data/ABS,LF,1.0.0/",
                 "M9.3.1599.30+10+20.AUS.M?startPeriod=2021-01&",
                 "endPeriod=2021-12&dimensionAtObservation=AllDimensions")

lf <- read_abs_api(lf_url)



# A URL that might need breaking up (note the different start date)
# (this URL doesn't actually need breaking up but it is a good example)

lf_url2 <- paste0("https://api.data.abs.gov.au/data/ABS,LF,1.0.0/",
                  "M9.3.1599.30+10+20.AUS.M?startPeriod=1978-01&",
                  "endPeriod=2021-12&dimensionAtObservation=AllDimensions")

lf_urls <- chunk_query_url(lf_url2)
read_api_safely <- purrr::safely(read_abs_api)
lf_raw <- purrr::map(lf_urls, read_api_safely)
lf_data <- purrr::map_dfr(lf_raw, "result")
}
}
