% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_sdmx.R
\name{read_abs_sdmx}
\alias{read_abs_sdmx}
\title{Extracts ABS XML-formatted data using the exported XML link at http://www.abs.gov.au/}
\usage{
read_abs_sdmx(url)
}
\arguments{
\item{url}{URL weblink.}
}
\value{
Dataframe
}
\description{
Extracts ABS XML-formatted data using the exported XML link at http://www.abs.gov.au/
}
