% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_sdmx.R
\name{read_abs_sdmx}
\alias{read_abs_sdmx}
\title{Extracts ABS XML-formatted data using the exported XML link at http://www.abs.gov.au/}
\usage{
read_abs_sdmx(url)
}
\arguments{
\item{url}{URL weblink.}
}
\value{
data frame
}
\description{
Access the sdmx URLs at http://www.abs.gov.au/ausstats/abs@.nsf/Lookup/1407.0.55.002Main+Features4User+Guide.
}
\examples{
\dontrun{
read_abs_sdmx("http://stat.data.abs.gov.au/restsdmx/sdmx.ashx/GetData/ABS_REGIONAL_ASGS/PENSION_2+BANKRUPT_2.AUS.0.A/all?startTime=2013&endTime=2013")
}
}
