% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_awe.R
\name{read_awe}
\alias{read_awe}
\title{read_awe}
\usage{
read_awe(
  wage_measure = c("awote", "ftawe", "awe"),
  sex = c("persons", "males", "females"),
  sector = c("total", "private", "public"),
  na.rm = FALSE,
  path = Sys.getenv("R_READABS_PATH", unset = tempdir()),
  show_progress_bars = FALSE,
  check_local = FALSE
)
}
\arguments{
\item{wage_measure}{Character of length 1. Must be one of:
\itemize{
  \item{`awote`}{ Average weekly ordinary time earnings; also known as Full-time adult ordinary time earnings}
  \item{`ftawe`}{ Full-time adult total earnings}
  \item{`awe`}{ Average weekly total earnings of all employees}
}}

\item{sex}{Character of length 1. Must be one of: `persons`, `males`, or `females`.}

\item{sector}{Character of length 1. Must be one of: `total`, `private`, or
`public`.}

\item{na.rm}{Logical. `FALSE` by default. If `FALSE`, a consistent quarterly
series is returned, with `NA` values for quarters in which there is no data.
If `TRUE`, only dates with data are included in the returned data frame.}

\item{path}{See `?read_abs`}

\item{show_progress_bars}{See `?read_abs`}

\item{check_local}{See `?read_abs`}
}
\value{
A `tbl_df` with four columns: `date`, `sex`, `wage_measure` and `value`.
The data is nominal (ie. not inflation-adjusted).
}
\description{
Convenience function to obtain wage levels from ABS
6302.0, Average Weekly Earnings, Australia.
}
\details{
The latest AWE data is available using `read_abs(cat_no = "6302.0", tables = 2)`.
However, this time series only goes back to 2012, when the ABS switched
from quarterly to biannual collection and release of the AWE data. The
`read_awe()` function assembles on time series back to November 1983 quarter;
it is quarterly to 2012 and biannual from then. Note that the data
returned with this function is consistently quarterly; any quarters for
which there are no observations are recorded as `NA` unless `na.rm` = `TRUE`.
}
\examples{
\dontrun{
read_awe("awote", "persons")
}

}
