% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_abs_catalogues.r
\name{scrape_abs_catalogues}
\alias{scrape_abs_catalogues}
\title{Helper function for \code{download_abs_data_cube} to scrape the available catalogues from the ABS website.}
\usage{
scrape_abs_catalogues()
}
\value{
A tibble containing the catalogues and how they are organised on the ABS website.
}
\description{
This function downloads a new version of the lookup table used by \code{show_available_catalogues}.
}
