% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_catalogues.r
\name{show_available_catalogues}
\alias{show_available_catalogues}
\title{Helper function for \code{download_abs_data_cube} to show the available catalogues.}
\usage{
show_available_catalogues(selected_heading = NULL)
}
\arguments{
\item{selected_heading}{optional character string specifying the heading on the \href{https://www.abs.gov.au/statistics}{ABS statistics webpage}.
e.g. "Earnings and work hours"}
}
\value{
a character vector of catalogues.
}
\description{
This function lists the possible catalogues that are available on the ABS website.
These catalogues must be specified as a string as an argument to \code{download_abs_data_cube}.
}
\examples{
show_available_catalogues("Earnings and work hours")

}
