% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_abs.R
\name{tidy_abs}
\alias{tidy_abs}
\title{Tidy ABS time series data.}
\usage{
tidy_abs(df)
}
\arguments{
\item{df}{A data frame containing ABS time series data that has been extracted
using \code{extract_abs_sheets}.}
}
\description{
Tidy ABS time series data.
}
\examples{

# First extract the data from the local spreadsheet

\dontrun{wpi <- extract_abs_sheets("634501.xls")}

# Then tidy the data extracted from the spreadsheet. Note that
# \\code{extract_abs_sheets()} returns a list of data frames, so we need to
# subset the list.

\donttest{tidy_wpi <- tidy_abs(wpi[[1]])}

}
