% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_api_data.R
\name{tidy_api_data}
\alias{tidy_api_data}
\title{Tidy the raw data returned using the ABS' API}
\usage{
tidy_api_data(url, old_api)
}
\arguments{
\item{url}{(char) The API URL used to query the data you want from the ABS.}

\item{old_api}{(boolean) States whether the ABS' API uses the old (ABS.Stat)
or new structure. If \code{old_api} == TRUE, then the API follows the old
structure.}
}
\value{
(tibble) Returns a tibble containing the ABS data you queried.
}
\description{
This function gets the nested list of raw data returned using the API URL and
transforms it into into a tidy tibble based on the structure of the API URL
(i.e. whether it follows the old or new structure, as stated in the
\code{old_api} parameter).
}
